projects/social-subtext-decoder/social-subtext-decoder/backend/requirements.txt
flask==2.3.3
deepface==0.0.90
opencv-python==4.10.0.84
pillow==10.0.0
numpy==1.26.4
